requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.5.1
pandas==2.2.3
numpy==1.26.4
xgboost==2.1.0
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
streamlit==1.39.0
requests==2.32.3
python-dotenv==1.0.1
